rays: Move OpenGL code into src/opengl/ and introduce Renderer abstraction
- Move opengl.h/cpp to src/opengl/
- Move platform-specific GL init ({osx,ios,sdl,win32}/opengl) to src/opengl/{plat}/
- Add src/renderer.h with Renderer_init/Renderer_fin abstraction
- Rename OpenGL_init/OpenGL_fin to Renderer_init/Renderer_fin
- Platform rays.{mm,cpp} now depends on renderer.h, not opengl.h
- Update all #include "opengl.h" to "opengl/opengl.h"
