+++ rays/src/color_space.h
@@ -5,7 +5,7 @@
 
 
 #include "rays/color_space.h"
-#include "opengl.h"
+#include "opengl/opengl.h"
 
 
 namespace Rays
