+++ rays/src/opengl/win32/opengl.cpp
@@ -1,7 +1,8 @@
-#include "../opengl.h"
+#include "../../renderer.h"
 
 
 #include <xot/windows.h>
+#include "../opengl.h"
 #include "rays/rays.h"
 #include "rays/exception.h"
 
@@ -90,7 +91,7 @@ namespace Rays
 	}
 
 	void
-	OpenGL_init ()
+	Renderer_init ()
 	{
 		activate_offscreen_context();
 
@@ -104,7 +105,7 @@ namespace Rays
 	}
 
 	void
-	OpenGL_fin ()
+	Renderer_fin ()
 	{
 	}
