+++ rays/src/renderer.h
@@ -1,0 +1,22 @@
+// -*- c++ -*-
+#pragma once
+#ifndef __RAYS_SRC_RENDERER_H__
+#define __RAYS_SRC_RENDERER_H__
+
+
+#include "rays/defs.h"
+
+
+namespace Rays
+{
+
+
+	void Renderer_init ();
+
+	void Renderer_fin ();
+
+
+}// Rays
+
+
+#endif//EOH
