+++ rays/src/shader.cpp
@@ -3,7 +3,7 @@
 
 #include <assert.h>
 #include "rays/exception.h"
-#include "opengl.h"
+#include "opengl/opengl.h"
 #include "image.h"
 #include "shader_program.h"
 #include "shader_source.h"
