+++ rays/src/shader_program.h
@@ -7,7 +7,7 @@
 #include <xot/pimpl.h>
 #include "rays/defs.h"
 #include "rays/coord.h"
-#include "opengl.h"
+#include "opengl/opengl.h"
 
 
 namespace Rays
